Island: Remove attack config and enable all techniques by default
