+++ monkey/monkey_island/cc/services/attack/attack_schema.py
@@ -10,8 +10,6 @@ SCHEMA = {
                 "T1059": {
                     "title": "Command line interface",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1059",
                     "description": "Adversaries may use command-line interfaces to interact with "
                     "systems "
@@ -20,8 +18,6 @@ SCHEMA = {
                 "T1106": {
                     "title": "Execution through API",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1106",
                     "description": "Adversary tools may directly use the Windows application "
                     "programming interface (API) to execute binaries.",
@@ -30,8 +26,6 @@ SCHEMA = {
                 "T1086": {
                     "title": "PowerShell",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1086",
                     "description": "Adversaries can use PowerShell to perform a number of actions,"
                     " including discovery of information and execution of code.",
@@ -39,8 +33,6 @@ SCHEMA = {
                 "T1064": {
                     "title": "Scripting",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1064",
                     "description": "Adversaries may use scripts to aid in operations and "
                     "perform multiple actions that would otherwise be manual.",
@@ -48,8 +40,6 @@ SCHEMA = {
                 "T1035": {
                     "title": "Service execution",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1035",
                     "description": "Adversaries may execute a binary, command, or script via a "
                     "method "
@@ -60,8 +50,6 @@ SCHEMA = {
                 "T1154": {
                     "title": "Trap",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1154",
                     "description": "Adversaries can use the trap command to register code to be "
                     "executed "
@@ -77,8 +65,6 @@ SCHEMA = {
                 "T1156": {
                     "title": ".bash_profile and .bashrc",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1156",
                     "description": "Adversaries may abuse shell scripts by "
                     "inserting arbitrary shell commands to gain persistence, which "
@@ -89,8 +75,6 @@ SCHEMA = {
                 "T1136": {
                     "title": "Create account",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1136",
                     "description": "Adversaries with a sufficient level of access "
                     "may create a local system, domain, or cloud tenant account.",
@@ -98,8 +82,6 @@ SCHEMA = {
                 "T1158": {
                     "title": "Hidden files and directories",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1158",
                     "description": "Adversaries can hide files and folders on the system "
                     "and evade a typical user or system analysis that does not "
@@ -108,8 +90,6 @@ SCHEMA = {
                 "T1168": {
                     "title": "Local job scheduling",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1168/",
                     "description": "Linux supports multiple methods for creating pre-scheduled and "
                     "periodic background jobs. Job scheduling can be used by "
@@ -121,8 +101,6 @@ SCHEMA = {
                 "T1504": {
                     "title": "PowerShell profile",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1504",
                     "description": "Adversaries may gain persistence and elevate privileges "
                     "in certain situations by abusing PowerShell profiles which "
@@ -132,8 +110,6 @@ SCHEMA = {
                 "T1053": {
                     "title": "Scheduled task",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1053",
                     "description": "Windows utilities can be used to schedule programs or scripts "
                     "to "
@@ -146,8 +122,6 @@ SCHEMA = {
                 "T1166": {
                     "title": "Setuid and Setgid",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1166",
                     "description": "Adversaries can set the setuid or setgid bits to get code "
                     "running in "
@@ -163,8 +137,6 @@ SCHEMA = {
                 "T1197": {
                     "title": "BITS jobs",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1197",
                     "description": "Adversaries may abuse BITS to download, execute, "
                     "and even clean up after running malicious code.",
@@ -172,8 +144,6 @@ SCHEMA = {
                 "T1146": {
                     "title": "Clear command history",
                     "type": "bool",
-                    "value": False,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1146",
                     "description": "Adversaries may clear/disable command history of a compromised "
                     "account to conceal the actions undertaken during an intrusion.",
@@ -181,8 +151,6 @@ SCHEMA = {
                 "T1107": {
                     "title": "File Deletion",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1107",
                     "description": "Adversaries may remove files over the course of an intrusion "
                     "to keep their footprint low or remove them at the end as part "
@@ -191,8 +159,6 @@ SCHEMA = {
                 "T1222": {
                     "title": "File permissions modification",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1222",
                     "description": "Adversaries may modify file permissions/attributes to evade "
                     "intended DACLs.",
@@ -200,8 +166,6 @@ SCHEMA = {
                 "T1099": {
                     "title": "Timestomping",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1099",
                     "description": "Adversaries may modify file time attributes to hide "
                     "new/changes to existing "
@@ -211,8 +175,6 @@ SCHEMA = {
                 "T1216": {
                     "title": "Signed script proxy execution",
                     "type": "bool",
-                    "value": False,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1216",
                     "description": "Adversaries may use scripts signed with trusted certificates "
                     "to proxy execution of malicious files on Windows systems. This behavior could "
@@ -229,8 +191,6 @@ SCHEMA = {
                 "T1110": {
                     "title": "Brute force",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1110",
                     "description": "Adversaries may use brute force techniques to attempt access "
                     "to accounts "
@@ -241,8 +201,6 @@ SCHEMA = {
                 "T1003": {
                     "title": "Credential dumping",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1003",
                     "description": "Mapped with T1078 Valid Accounts because "
                     "both techniques require"
@@ -257,8 +215,6 @@ SCHEMA = {
                 "T1145": {
                     "title": "Private keys",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1145",
                     "description": "Adversaries may gather private keys from compromised systems "
                     "for use in "
@@ -277,8 +233,6 @@ SCHEMA = {
                 "T1087": {
                     "title": "Account Discovery",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1087",
                     "description": "Adversaries may attempt to get a listing of accounts on a "
                     "system or "
@@ -289,8 +243,6 @@ SCHEMA = {
                 "T1018": {
                     "title": "Remote System Discovery",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1018",
                     "description": "Adversaries will likely attempt to get a listing of other "
                     "systems by IP address, "
@@ -300,8 +252,6 @@ SCHEMA = {
                 "T1082": {
                     "title": "System information discovery",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1082",
                     "depends_on": ["T1016", "T1005"],
                     "description": "An adversary may attempt to get detailed information about the "
@@ -312,8 +262,6 @@ SCHEMA = {
                 "T1016": {
                     "title": "System network configuration discovery",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1016",
                     "depends_on": ["T1005", "T1082"],
                     "description": "Adversaries will likely look for details about the network "
@@ -332,8 +280,6 @@ SCHEMA = {
                 "T1210": {
                     "title": "Exploitation of Remote services",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1210",
                     "description": "Exploitation of a software vulnerability occurs when an "
                     "adversary "
@@ -345,8 +291,6 @@ SCHEMA = {
                 "T1075": {
                     "title": "Pass the hash",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1075",
                     "description": "Pass the hash (PtH) is a method of authenticating as a user "
                     "without "
@@ -355,8 +299,6 @@ SCHEMA = {
                 "T1105": {
                     "title": "Remote file copy",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1105",
                     "description": "Files may be copied from one system to another to stage "
                     "adversary tools or other files over the course of an operation.",
@@ -364,8 +306,6 @@ SCHEMA = {
                 "T1021": {
                     "title": "Remote services",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1021",
                     "depends_on": ["T1110"],
                     "description": "An adversary may use Valid Accounts to log into a service"
@@ -381,8 +321,6 @@ SCHEMA = {
                 "T1005": {
                     "title": "Data from local system",
                     "type": "bool",
-                    "value": True,
-                    "necessary": False,
                     "link": "https://attack.mitre.org/techniques/T1005",
                     "depends_on": ["T1016", "T1082"],
                     "description": "Sensitive data can be collected from local system sources, "
@@ -400,8 +338,6 @@ SCHEMA = {
                 "T1090": {
                     "title": "Connection proxy",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1090",
                     "description": "A connection proxy is used to direct network traffic between "
                     "systems "
@@ -410,8 +346,6 @@ SCHEMA = {
                 "T1065": {
                     "title": "Uncommonly used port",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1065",
                     "description": "Adversaries may conduct C2 communications over a non-standard "
                     "port to bypass proxies and firewalls that have been improperly "
@@ -420,8 +354,6 @@ SCHEMA = {
                 "T1188": {
                     "title": "Multi-hop proxy",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1188",
                     "description": "To disguise the source of malicious traffic, "
                     "adversaries may chain together multiple proxies.",
@@ -436,8 +368,6 @@ SCHEMA = {
                 "T1041": {
                     "title": "Exfiltration Over Command and Control Channel",
                     "type": "bool",
-                    "value": True,
-                    "necessary": True,
                     "link": "https://attack.mitre.org/techniques/T1041",
                     "description": "Data exfiltration is performed over the Command and Control "
                     "channel.",
