+++ monkey/monkey_island/cc/services/attack/technique_reports/T1210.py
@@ -16,7 +16,6 @@ class T1210(AttackTechnique):
 
     @staticmethod
     def get_report_data():
-        @T1210.is_status_disabled
         def get_technique_status_and_data():
             scanned_services = T1210.get_scanned_services()
             exploited_services = T1210.get_exploited_services()
@@ -30,10 +29,8 @@ class T1210(AttackTechnique):
 
         status_and_data = get_technique_status_and_data()
         status = status_and_data[0]
-        if status == ScanStatus.DISABLED.value:
-            scanned_services, exploited_services = [], []
-        else:
-            scanned_services, exploited_services = status_and_data[1], status_and_data[2]
+
+        scanned_services, exploited_services = status_and_data[1], status_and_data[2]
         data = {"title": T1210.technique_title()}
 
         data.update(T1210.get_message_and_status(status))
